requirements.txt
django
requests
PyJWT
